backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Marshmallow==1.2.1
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
flask-cors
